requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.1
supabase==2.3.1
python-multipart==0.0.9
mistralai==0.1.3
pillow==10.2.0
pdf2image==1.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4